COMP: Fixed the warning.
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=7239
git-svn-id: http://svn.slicer.org/Slicer4/trunk@7239 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Libs/MRML/vtkMRMLGridTransformNode.cxx
@@ -169,10 +169,10 @@ void vtkMRMLGridTransformNode::ReadXMLAttributes(const char** atts)
       }
     else if (!strcmp(attName, "dimension"))
       {
-      double val;
+      int val;
       std::stringstream ss;
       ss << attValue;
-      std::vector<double> vals;
+      std::vector<int> vals;
       while( ss >> val )
         {
         vals.push_back( val );
